requirements.txt
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.5
Brlapi==0.8.4
bs4==0.0.2
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
distro==1.8.0
distro-info==1.5+deb12u1
duplicity==0.8.22
fasteners==0.17.3
future==0.18.2
httplib2==0.20.4
idna==3.3
invoke==2.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.24.0
Mako==1.2.4.dev0
mariadb==1.1.10
MarkupSafe==2.1.2
monotonic==1.6
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.4.0
psutil==5.9.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt6==6.4.2
PyQt6-sip==13.4.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-pam==2.0.2
python-xlib==0.33
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
reportbug==12.0.0
requests==2.28.1
SCons==4.4.0
setproctitle==1.3.1
six==1.16.0
soupsieve==2.5
tinycss==0.4
tinycss2==1.2.1
urllib3==1.26.12
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xlrd==1.2.0
